+++ src/main/java/com/bachar/customer/CustomerRepository.java
@@ -2,9 +2,9 @@ package com.bachar.customer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
     boolean existsCustomerByEmail(String email);
 
-    boolean existsCustomerById(Integer customerId);
+    boolean existsCustomerById(Long customerId);
 }
